Rename js files to use kebab-case
